Change setting namespaces on resources.
This allows creating resources in other namespaces.

It does this by no longer forcing the resources into the template
namespace.

If the rendered resource does not provide a namespace, it will get the
template namespace - we need this to create the resource in the API.

It adds a top-level `.Template.{Name,Namespace} to make it easy to get
the namespace of the GitOpsSet if you want to explicitly declare it.