Suppress caching of Secrets and ConfigMaps
As explained in
https://github.com/kubernetes-sigs/controller-runtime/pull/1249, the
controller-runtime client will tend to use LIST and WATCH to cache
resources requested by the controller. This has two downsides:

 - the mechanism may require broad RBAC access (Secret and ConfigMap
   in every namespace) even if the actual secrets and configmaps
   referenced are few, or restricted to few namespaces;
 - the cache uses a lot of memory that it doesn't really need to

This change fixes those problems, with the trade-off that all Secret
and ConfigMap requests use a round-trip to the Kubernetes API server.

This follows
https://github.com/fluxcd/source-controller/pull/989. There, a feature
flag can be used to restore caching for Secrets and ConfigMaps; I have
not included the feature flag, since this controller is much younger
and doesn't have an established behaviour.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>