Fix test flake due to expecting exactly one event
The test TestEventsWithReconciling checks that a resource has an event
indicating a successful reconciliation, when reconciliation
succeeds. It can be flakey, though, because it asserts the value of
the whole list of events, and the controller can post multiple events
before it gets checked.

So, instead: check that the list of events _contains_ an event
indicating success.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>