Wait for Artifact before generation.
The artifact may or may not be available, if we've already generated
resources, generating no resources means that the existing resources
are removed.

This changes it to return a marker error, which we can record and do
nothing about.

We can do nothing, because when the Artifact is applied, this will
trigger a watch notification, and we can generate.